+++ manifest-editor/src/editors/CanvasProperties/AnnotationForm.tsx
@@ -2,9 +2,7 @@ import { IIIFBuilder } from "iiif-builder";
 import { useCanvas, useVault } from "react-iiif-vault";
 import { useManifest } from "../../hooks/useManifest";
 import { EmptyProperty } from "../../atoms/EmptyProperty";
-import { InformationLink } from "../../atoms/InformationLink";
-import { FlexContainer } from "../../components/layout/FlexContainer";
-import { AnnotationPreview } from "./AnnotationPreview";
+import { AnnotationPreview } from "./Annotation";
 import { v4 } from "uuid";
 
 export const AnnotationForm = () => {
@@ -40,21 +38,7 @@ export const AnnotationForm = () => {
   };
   return (
     <>
-      <EmptyProperty label={"annotations"} createNew={addNew} />
-      <div>
-        <FlexContainer>
-          <h4>annotations</h4>
-          {guidanceReference && <InformationLink guidanceReference={guidanceReference} />}
-        </FlexContainer>
-        <pre
-          // @ts-ignore
-          dangerouslySetInnerHTML={{
-            // @ts-ignore
-            __html: JSON.stringify(vault.get(canvas?.annotations), null, 3),
-          }}
-        ></pre>
-      </div>
-      {/* <h4>{vault.get(canvas.annotations)[0].items.length}</h4> */}
+      <EmptyProperty label={"annotations"} createNew={addNew} guidanceReference={guidanceReference} />
       {vault &&
         canvas &&
         // @ts-ignore
@@ -64,7 +48,7 @@ export const AnnotationForm = () => {
           // @ts-ignore
           return vault.get(item.id)?.items.map((NESTEDITEM: any) => {
             // console.log(NESTEDITEM);
-            return <AnnotationPreview thumbnailSrc={NESTEDITEM.id} />;
+            return <AnnotationPreview id={NESTEDITEM.id} />;
           });
         })}
     </>
